fix:Make Navbar Sticky with Bottom Border and Shadow
